Minor fixes (#33)
* Minor fixes

* export getProjectRoot

* .
